requirements.txt
Flask==1.1.2
gunicorn==20.1.0
tensorflow-cpu
keras 
numpy